+++ docs/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Next
+
 * Added back in Repuation for D2.
+* Max Light Loadout, Make Room for Postmaster, Farming Mode, and Search Loadout are all reenabled for D2.
 
 # 4.13.0